+++ stm32-smart-inhaler/Core/Inc/rtc.h
@@ -1,0 +1,16 @@
+/*
+ * rtc.h
+ *
+ *  Created on: Jun 14, 2021
+ *      Author: youss
+ */
+
+#ifndef CORE_INC_RTC_H_
+#define CORE_INC_RTC_H_
+
+#include <time.h>
+#include "stm32wbxx_hal.h"
+
+uint32_t get_timestamp();
+
+#endif /* CORE_INC_RTC_H_ */
